proxygen: share one source tree, preload static assets, cut per-request cost
Builds on the two Proxygen engine commits. Both entries now build from
frameworks/proxygen/src selected by a TARGET build arg, the way sark-h3 reuses
sark, so a fix lands in both entries at once.

Static assets and their precompressed .br/.gz siblings are read once at startup
and served as non-owning IOBuf views: no disk I/O, no per-request compression,
no copy, no payload allocation. This replaces an open/read plus an on-the-fly
gzip of every CSS/JS/HTML response, which dominated the static profiles.

Also:
- shared RFC 6455 codec: frames unmasked in place, one egress write per read
- response compression scoped to application/json, the type json-comp scores
- coro attaches CompressionFilter per response instead of registering a
  server-level factory, which allocated four objects on every request because
  HTTPFilterFactoryHandler builds filters before reading request headers
- per-request allocations removed from the hot path: getPathAsStringPiece, a
  query-string scan instead of the parsed parameter map, string_view content
  types, to_chars, and no shared_ptr copy per handler
- mimalloc preloaded; separate --threads / --h3_threads
- base image pinned by digest for reproducible rounds

Measured locally on a 32-core cpuset with wrk, so relative deltas only:

  static      13,012 -> 1,049,630 rps    static-tls  12,701 -> 706,558 rps
  json        +38%                       json-comp   +41% classic / +82% coro
  baseline    +7.5% classic / +17% coro  websocket   -30% / -46% CPU per frame

Peak memory on static at 6800 connections drops 2.1 GiB -> 160 MiB.

validate.sh reports 63/63 for both entries; HTTP/3 and RFC 6455 conformance
checked separately.
