+++ frameworks/proxygen/entrypoint.sh
@@ -1,7 +1,12 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-exec /usr/local/bin/proxygen-arena \
+# Shared by the `proxygen` and `proxygen-coro` images; the binary is the same
+# name in both, only the server API compiled into it differs.
+#
+# PROXYGEN_THREADS    TCP (H1/H2) I/O threads, 0 = available CPUs
+# PROXYGEN_H3_THREADS QUIC I/O threads, 0 = available CPUs
+exec /usr/local/bin/arena-server \
     --ip=:: \
     --http_port=8080 \
     --tls_port=8081 \
@@ -10,4 +15,5 @@ exec /usr/local/bin/proxygen-arena \
     --h3_port=8443 \
     --cert=/certs/server.crt \
     --key=/certs/server.key \
-    --threads="${PROXYGEN_THREADS:-0}"
+    --threads="${PROXYGEN_THREADS:-0}" \
+    --h3_threads="${PROXYGEN_H3_THREADS:-0}"
